+++ raiden-dapp/src/services/raiden-service.ts
@@ -1,4 +1,5 @@
 import {
+  Address,
   Capabilities,
   ChangeEvent,
   ErrorCodes,
@@ -669,7 +670,11 @@ export default class RaidenService {
     if (!raiden.mainAddress) {
       return [];
     }
-    const allTokens = await raiden.getTokenList();
+    const allTokenAddresses = uniq(
+      (await raiden.getTokenList()).concat(
+        this._userDepositTokenAddress as Address
+      )
+    );
     const balances: Tokens = {};
     const fetchTokenBalance = async (address: string): Promise<void> =>
       raiden.getTokenBalance(address, raiden.address).then((balance) => {
@@ -681,7 +686,7 @@ export default class RaidenService {
         }
       });
 
-    await asyncPool(6, allTokens, fetchTokenBalance);
+    await asyncPool(6, allTokenAddresses, fetchTokenBalance);
     const tokens: Tokens = {};
     Object.keys(balances).forEach((address) => {
       const cached = this.store.state.tokens[address];